+++ programs/drift/src/instructions/admin.rs
@@ -267,7 +267,8 @@ pub fn handle_initialize_spot_market(
         flash_loan_amount: 0,
         flash_loan_initial_token_amount: 0,
         total_swap_fee: 0,
-        padding: [0; 56],
+        scale_initial_asset_weight_start: 0,
+        padding: [0; 48],
         insurance_fund: InsuranceFund {
             vault: *ctx.accounts.insurance_fund_vault.to_account_info().key,
             unstaking_period: THIRTEEN_DAY,
@@ -1690,6 +1691,18 @@ pub fn handle_update_spot_market_max_token_deposits(
 ) -> Result<()> {
     let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
     spot_market.max_token_deposits = max_token_deposits;
+    Ok(())
+}
+
+#[access_control(
+spot_market_valid(&ctx.accounts.spot_market)
+)]
+pub fn handle_update_spot_market_scale_initial_asset_weight_start(
+    ctx: Context<AdminUpdateSpotMarket>,
+    scale_initial_asset_weight_start: u64,
+) -> Result<()> {
+    let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
+    spot_market.scale_initial_asset_weight_start = scale_initial_asset_weight_start;
     Ok(())
 }