+++ programs/drift/src/math/spot_swap.rs
@@ -2,7 +2,8 @@ use crate::error::DriftResult;
 use crate::math::casting::Cast;
 use crate::math::margin::MarginRequirementType;
 use crate::math::safe_math::SafeMath;
-use crate::math::spot_balance::get_token_value;
+use crate::math::spot_balance::get_strict_token_value;
+use crate::state::oracle::StrictOraclePrice;
 use crate::state::spot_market::SpotMarket;
 use crate::{PRICE_PRECISION, SPOT_WEIGHT_PRECISION_U128};
 
@@ -25,8 +26,8 @@ pub fn calculate_swap_price(
 pub fn select_margin_type_for_swap(
     in_market: &SpotMarket,
     out_market: &SpotMarket,
-    in_price: i64,
-    out_price: i64,
+    in_strict_price: &StrictOraclePrice,
+    out_strict_price: &StrictOraclePrice,
     in_token_amount_before: i128,
     out_token_amount_before: i128,
     in_token_amount_after: i128,
@@ -34,12 +35,14 @@ pub fn select_margin_type_for_swap(
     strict_margin_type: MarginRequirementType,
 ) -> DriftResult<MarginRequirementType> {
     let calculate_free_collateral_contribution =
-        |market: &SpotMarket, price: i64, token_amount: i128| {
-            let token_value = get_token_value(token_amount, market.decimals, price)?;
+        |market: &SpotMarket, strict_oracle_price: &StrictOraclePrice, token_amount: i128| {
+            let token_value =
+                get_strict_token_value(token_amount, market.decimals, strict_oracle_price)?;
 
             let weight = if token_amount >= 0 {
                 market.get_asset_weight(
                     token_amount.unsigned_abs(),
+                    strict_oracle_price.current,
                     &MarginRequirementType::Initial,
                 )?
             } else {
@@ -55,19 +58,25 @@ pub fn select_margin_type_for_swap(
         };
 
     let in_free_collateral_contribution_before =
-        calculate_free_collateral_contribution(in_market, in_price, in_token_amount_before)?;
+        calculate_free_collateral_contribution(in_market, in_strict_price, in_token_amount_before)?;
 
-    let out_free_collateral_contribution_before =
-        calculate_free_collateral_contribution(out_market, out_price, out_token_amount_before)?;
+    let out_free_collateral_contribution_before = calculate_free_collateral_contribution(
+        out_market,
+        out_strict_price,
+        out_token_amount_before,
+    )?;
 
     let free_collateral_contribution_before =
         in_free_collateral_contribution_before.safe_add(out_free_collateral_contribution_before)?;
 
     let in_free_collateral_contribution_after =
-        calculate_free_collateral_contribution(in_market, in_price, in_token_amount_after)?;
+        calculate_free_collateral_contribution(in_market, in_strict_price, in_token_amount_after)?;
 
-    let out_free_collateral_contribution_after =
-        calculate_free_collateral_contribution(out_market, out_price, out_token_amount_after)?;
+    let out_free_collateral_contribution_after = calculate_free_collateral_contribution(
+        out_market,
+        out_strict_price,
+        out_token_amount_after,
+    )?;
 
     let free_collateral_contribution_after =
         in_free_collateral_contribution_after.safe_add(out_free_collateral_contribution_after)?;