+++ sdk/src/user.ts
@@ -73,12 +73,14 @@ import { UserConfig } from './userConfig';
 import { PollingUserAccountSubscriber } from './accounts/pollingUserAccountSubscriber';
 import { WebSocketUserAccountSubscriber } from './accounts/webSocketUserAccountSubscriber';
 import {
+	calculateWeightedTokenValue,
 	getWorstCaseTokenAmounts,
 	isSpotPositionAvailable,
 } from './math/spotPosition';
 
 import { calculateLiveOracleTwap } from './math/oracles';
 import { getPerpMarketTierNumber, getSpotMarketTierNumber } from './math/tiers';
+import { StrictOraclePrice } from './oracles/strictOraclePrice';
 
 export class User {
 	driftClient: DriftClient;
@@ -869,6 +871,20 @@ export class User {
 				spotPosition.marketIndex
 			);
 
+			let twap5min;
+			if (strict) {
+				twap5min = calculateLiveOracleTwap(
+					spotMarketAccount.historicalOracleData,
+					oraclePriceData,
+					now,
+					FIVE_MINUTE // 5MIN
+				);
+			}
+			const strictOraclePrice = new StrictOraclePrice(
+				oraclePriceData.price,
+				twap5min
+			);
+
 			if (
 				spotPosition.marketIndex === QUOTE_SPOT_MARKET_INDEX &&
 				countForQuote
@@ -885,23 +901,19 @@ export class User {
 				if (isVariant(spotPosition.balanceType, 'borrow')) {
 					const weightedTokenValue = this.getSpotLiabilityValue(
 						tokenAmount,
-						oraclePriceData,
+						strictOraclePrice,
 						spotMarketAccount,
 						marginCategory,
-						liquidationBuffer,
-						strict,
-						now
+						liquidationBuffer
 					).abs();
 
 					netQuoteValue = netQuoteValue.sub(weightedTokenValue);
 				} else {
 					const weightedTokenValue = this.getSpotAssetValue(
 						tokenAmount,
-						oraclePriceData,
+						strictOraclePrice,
 						spotMarketAccount,
-						marginCategory,
-						strict,
-						now
+						marginCategory
 					);
 
 					netQuoteValue = netQuoteValue.add(weightedTokenValue);
@@ -922,12 +934,10 @@ export class User {
 					);
 					const liabilityValue = this.getSpotLiabilityValue(
 						tokenAmount,
-						oraclePriceData,
+						strictOraclePrice,
 						spotMarketAccount,
 						marginCategory,
-						liquidationBuffer,
-						strict,
-						now
+						liquidationBuffer
 					).abs();
 					totalLiabilityValue = totalLiabilityValue.add(liabilityValue);
 
@@ -940,11 +950,9 @@ export class User {
 					);
 					const assetValue = this.getSpotAssetValue(
 						tokenAmount,
-						oraclePriceData,
+						strictOraclePrice,
 						spotMarketAccount,
-						marginCategory,
-						strict,
-						now
+						marginCategory
 					);
 					totalAssetValue = totalAssetValue.add(assetValue);
 
@@ -952,21 +960,22 @@ export class User {
 				}
 			}
 
-			const [worstCaseTokenAmount, worstCaseQuoteTokenAmount] =
-				getWorstCaseTokenAmounts(
-					spotPosition,
-					spotMarketAccount,
-					this.getOracleDataForSpotMarket(spotPosition.marketIndex)
-				);
+			const {
+				tokenAmount: worstCaseTokenAmount,
+				ordersValue: worstCaseQuoteTokenAmount,
+			} = getWorstCaseTokenAmounts(
+				spotPosition,
+				spotMarketAccount,
+				strictOraclePrice,
+				marginCategory
+			);
 
 			if (worstCaseTokenAmount.gt(ZERO) && countForBase) {
 				const baseAssetValue = this.getSpotAssetValue(
 					worstCaseTokenAmount,
-					oraclePriceData,
+					strictOraclePrice,
 					spotMarketAccount,
-					marginCategory,
-					strict,
-					now
+					marginCategory
 				);
 
 				totalAssetValue = totalAssetValue.add(baseAssetValue);
@@ -975,12 +984,10 @@ export class User {
 			if (worstCaseTokenAmount.lt(ZERO) && countForBase) {
 				const baseLiabilityValue = this.getSpotLiabilityValue(
 					worstCaseTokenAmount,
-					oraclePriceData,
+					strictOraclePrice,
 					spotMarketAccount,
 					marginCategory,
-					liquidationBuffer,
-					strict,
-					now
+					liquidationBuffer
 				).abs();
 
 				totalLiabilityValue = totalLiabilityValue.add(baseLiabilityValue);
@@ -1041,35 +1048,16 @@ export class User {
 
 	getSpotLiabilityValue(
 		tokenAmount: BN,
-		oraclePriceData: OraclePriceData,
+		strictOraclePrice: StrictOraclePrice,
 		spotMarketAccount: SpotMarketAccount,
 		marginCategory?: MarginCategory,
-		liquidationBuffer?: BN,
-		strict = false,
-		now?: BN
+		liquidationBuffer?: BN
 	): BN {
-		let liabilityValue = null;
-
-		if (strict) {
-			const estOracleTwap = calculateLiveOracleTwap(
-				spotMarketAccount.historicalOracleData,
-				oraclePriceData,
-				now,
-				FIVE_MINUTE // 5MIN
-			);
-			liabilityValue = getStrictTokenValue(
-				tokenAmount,
-				spotMarketAccount.decimals,
-				oraclePriceData,
-				estOracleTwap
-			);
-		} else {
-			liabilityValue = getTokenValue(
-				tokenAmount,
-				spotMarketAccount.decimals,
-				oraclePriceData
-			);
-		}
+		let liabilityValue = getStrictTokenValue(
+			tokenAmount,
+			spotMarketAccount.decimals,
+			strictOraclePrice
+		);
 
 		if (marginCategory !== undefined) {
 			let weight = calculateLiabilityWeight(
@@ -1114,37 +1102,20 @@ export class User {
 
 	getSpotAssetValue(
 		tokenAmount: BN,
-		oraclePriceData: OraclePriceData,
+		strictOraclePrice: StrictOraclePrice,
 		spotMarketAccount: SpotMarketAccount,
-		marginCategory?: MarginCategory,
-		strict = false,
-		now?: BN
+		marginCategory?: MarginCategory
 	): BN {
-		let assetValue = null;
-		if (strict) {
-			const estOracleTwap = calculateLiveOracleTwap(
-				spotMarketAccount.historicalOracleData,
-				oraclePriceData,
-				now,
-				FIVE_MINUTE // 5MIN
-			);
-			assetValue = getStrictTokenValue(
-				tokenAmount,
-				spotMarketAccount.decimals,
-				oraclePriceData,
-				estOracleTwap
-			);
-		} else {
-			assetValue = getTokenValue(
-				tokenAmount,
-				spotMarketAccount.decimals,
-				oraclePriceData
-			);
-		}
+		let assetValue = getStrictTokenValue(
+			tokenAmount,
+			spotMarketAccount.decimals,
+			strictOraclePrice
+		);
 
 		if (marginCategory !== undefined) {
 			const weight = calculateAssetWeight(
 				tokenAmount,
+				strictOraclePrice.current,
 				spotMarketAccount,
 				marginCategory
 			);
@@ -2092,6 +2063,7 @@ export class User {
 		if (signedTokenAmount.gt(ZERO)) {
 			const assetWeight = calculateAssetWeight(
 				signedTokenAmount,
+				this.driftClient.getOraclePriceDataAndSlot(market.oracle).data.price,
 				market,
 				'Maintenance'
 			);
@@ -2268,6 +2240,9 @@ export class User {
 		currentSpotMarketNetValue?: BN
 	): BN {
 		const market = this.driftClient.getSpotMarketAccount(targetMarketIndex);
+		const oraclePrice = this.driftClient.getOraclePriceDataAndSlot(
+			market.oracle
+		).data.price;
 
 		currentQuoteAssetValue = this.getSpotMarketAssetValue(
 			QUOTE_SPOT_MARKET_INDEX
@@ -2279,6 +2254,7 @@ export class User {
 		let freeCollateral = this.getFreeCollateral();
 		const marginRatio = calculateSpotMarketMarginRatio(
 			market,
+			oraclePrice,
 			'Initial',
 			ZERO,
 			isVariant(direction, 'long')
@@ -2293,6 +2269,7 @@ export class User {
 				tradeAmount = currentSpotMarketNetValue.abs();
 				const marginRatio = calculateSpotMarketMarginRatio(
 					market,
+					oraclePrice,
 					'Initial',
 					this.getTokenAmount(targetMarketIndex).abs(),
 					SpotBalanceType.BORROW
@@ -2307,6 +2284,7 @@ export class User {
 				tradeAmount = currentSpotMarketNetValue;
 				const marginRatio = calculateSpotMarketMarginRatio(
 					market,
+					oraclePrice,
 					'Initial',
 					this.getTokenAmount(targetMarketIndex),
 					SpotBalanceType.DEPOSIT
@@ -2354,9 +2332,13 @@ export class User {
 		const inMarket = this.driftClient.getSpotMarketAccount(inMarketIndex);
 		const outMarket = this.driftClient.getSpotMarketAccount(outMarketIndex);
 
-		const inOraclePrice = this.getOracleDataForSpotMarket(inMarketIndex).price;
-		const outOraclePrice =
-			this.getOracleDataForSpotMarket(outMarketIndex).price;
+		const inOraclePriceData = this.getOracleDataForSpotMarket(inMarketIndex);
+		const inOraclePrice = inOraclePriceData.price;
+		const outOraclePriceData = this.getOracleDataForSpotMarket(outMarketIndex);
+		const outOraclePrice = outOraclePriceData.price;
+
+		const inStrictOraclePrice = new StrictOraclePrice(inOraclePrice);
+		const outStrictOraclePrice = new StrictOraclePrice(outOraclePrice);
 
 		const inPrecision = new BN(10 ** inMarket.decimals);
 		const outPrecision = new BN(10 ** outMarket.decimals);
@@ -2371,17 +2353,29 @@ export class User {
 		const freeCollateral = this.getFreeCollateral();
 
 		const inContributionInitial =
-			this.calculateSpotPositionFreeCollateralContribution(inSpotPosition);
+			this.calculateSpotPositionFreeCollateralContribution(
+				inSpotPosition,
+				inStrictOraclePrice
+			);
 		const {
 			totalAssetValue: inTotalAssetValueInitial,
 			totalLiabilityValue: inTotalLiabilityValueInitial,
-		} = this.calculateSpotPositionLeverageContribution(inSpotPosition);
+		} = this.calculateSpotPositionLeverageContribution(
+			inSpotPosition,
+			inStrictOraclePrice
+		);
 		const outContributionInitial =
-			this.calculateSpotPositionFreeCollateralContribution(outSpotPosition);
+			this.calculateSpotPositionFreeCollateralContribution(
+				outSpotPosition,
+				outStrictOraclePrice
+			);
 		const {
 			totalAssetValue: outTotalAssetValueInitial,
 			totalLiabilityValue: outTotalLiabilityValueInitial,
-		} = this.calculateSpotPositionLeverageContribution(outSpotPosition);
+		} = this.calculateSpotPositionLeverageContribution(
+			outSpotPosition,
+			outStrictOraclePrice
+		);
 		const initialContribution = inContributionInitial.add(
 			outContributionInitial
 		);
@@ -2441,10 +2435,14 @@ export class User {
 				);
 
 				const inContributionAfter =
-					this.calculateSpotPositionFreeCollateralContribution(inPositionAfter);
+					this.calculateSpotPositionFreeCollateralContribution(
+						inPositionAfter,
+						inStrictOraclePrice
+					);
 				const outContributionAfter =
 					this.calculateSpotPositionFreeCollateralContribution(
-						outPositionAfter
+						outPositionAfter,
+						outStrictOraclePrice
 					);
 
 				const contributionAfter = inContributionAfter.add(outContributionAfter);
@@ -2482,12 +2480,18 @@ export class User {
 		const {
 			totalAssetValue: inTotalAssetValueAfter,
 			totalLiabilityValue: inTotalLiabilityValueAfter,
-		} = this.calculateSpotPositionLeverageContribution(inPositionAfter);
+		} = this.calculateSpotPositionLeverageContribution(
+			inPositionAfter,
+			inStrictOraclePrice
+		);
 
 		const {
 			totalAssetValue: outTotalAssetValueAfter,
 			totalLiabilityValue: outTotalLiabilityValueAfter,
-		} = this.calculateSpotPositionLeverageContribution(outPositionAfter);
+		} = this.calculateSpotPositionLeverageContribution(
+			outPositionAfter,
+			outStrictOraclePrice
+		);
 
 		const spotAssetValueDelta = inTotalAssetValueAfter
 			.add(outTotalAssetValueAfter)
@@ -2564,108 +2568,54 @@ export class User {
 	}
 
 	calculateSpotPositionFreeCollateralContribution(
-		spotPosition: SpotPosition
+		spotPosition: SpotPosition,
+		strictOraclePrice: StrictOraclePrice
 	): BN {
-		let freeCollateralContribution = ZERO;
-		const now = new BN(new Date().getTime() / 1000);
-		const strict = true;
 		const marginCategory = 'Initial';
 
 		const spotMarketAccount: SpotMarketAccount =
 			this.driftClient.getSpotMarketAccount(spotPosition.marketIndex);
 
-		const oraclePriceData = this.getOracleDataForSpotMarket(
-			spotPosition.marketIndex
+		const { freeCollateralContribution } = getWorstCaseTokenAmounts(
+			spotPosition,
+			spotMarketAccount,
+			strictOraclePrice,
+			marginCategory
 		);
-
-		const [worstCaseTokenAmount, worstCaseQuoteTokenAmount] =
-			getWorstCaseTokenAmounts(
-				spotPosition,
-				spotMarketAccount,
-				oraclePriceData
-			);
-
-		if (worstCaseTokenAmount.gt(ZERO)) {
-			const baseAssetValue = this.getSpotAssetValue(
-				worstCaseTokenAmount,
-				oraclePriceData,
-				spotMarketAccount,
-				marginCategory,
-				strict,
-				now
-			);
-
-			freeCollateralContribution =
-				freeCollateralContribution.add(baseAssetValue);
-		} else {
-			const baseLiabilityValue = this.getSpotLiabilityValue(
-				worstCaseTokenAmount,
-				oraclePriceData,
-				spotMarketAccount,
-				marginCategory,
-				undefined,
-				strict,
-				now
-			).abs();
-
-			freeCollateralContribution =
-				freeCollateralContribution.sub(baseLiabilityValue);
-		}
-
-		freeCollateralContribution.add(worstCaseQuoteTokenAmount);
 
 		return freeCollateralContribution;
 	}
 
-	calculateSpotPositionLeverageContribution(spotPosition: SpotPosition): {
+	calculateSpotPositionLeverageContribution(
+		spotPosition: SpotPosition,
+		strictOraclePrice: StrictOraclePrice
+	): {
 		totalAssetValue: BN;
 		totalLiabilityValue: BN;
 	} {
 		let totalAssetValue = ZERO;
 		let totalLiabilityValue = ZERO;
-		const now = new BN(new Date().getTime() / 1000);
 
 		const spotMarketAccount: SpotMarketAccount =
 			this.driftClient.getSpotMarketAccount(spotPosition.marketIndex);
 
-		const oraclePriceData = this.getOracleDataForSpotMarket(
-			spotPosition.marketIndex
+		const { tokenValue, ordersValue } = getWorstCaseTokenAmounts(
+			spotPosition,
+			spotMarketAccount,
+			strictOraclePrice,
+			'Initial'
 		);
 
-		const [worstCaseTokenAmount, worstCaseQuoteTokenAmount] =
-			getWorstCaseTokenAmounts(
-				spotPosition,
-				spotMarketAccount,
-				oraclePriceData
-			);
-
-		if (worstCaseTokenAmount.gt(ZERO)) {
-			totalAssetValue = this.getSpotAssetValue(
-				worstCaseTokenAmount,
-				oraclePriceData,
-				spotMarketAccount,
-				undefined,
-				false,
-				now
-			);
+		if (tokenValue.gte(ZERO)) {
+			totalAssetValue = tokenValue;
 		} else {
-			totalLiabilityValue = this.getSpotLiabilityValue(
-				worstCaseTokenAmount,
-				oraclePriceData,
-				spotMarketAccount,
-				undefined,
-				undefined,
-				false,
-				now
-			).abs();
+			totalLiabilityValue = tokenValue.abs();
 		}
 
-		if (worstCaseQuoteTokenAmount.gt(ZERO)) {
-			totalAssetValue = totalAssetValue.add(worstCaseQuoteTokenAmount);
+		if (ordersValue.gt(ZERO)) {
+			totalAssetValue = totalAssetValue.add(ordersValue);
 		} else {
-			totalLiabilityValue = totalLiabilityValue.add(
-				worstCaseQuoteTokenAmount.abs()
-			);
+			totalLiabilityValue = totalLiabilityValue.add(ordersValue.abs());
 		}
 
 		return {
@@ -2695,6 +2645,13 @@ export class User {
 		const inMarket = this.driftClient.getSpotMarketAccount(inMarketIndex);
 		const outMarket = this.driftClient.getSpotMarketAccount(outMarketIndex);
 
+		const inOraclePriceData = this.getOracleDataForSpotMarket(inMarketIndex);
+		const inOraclePrice = inOraclePriceData.price;
+		const outOraclePriceData = this.getOracleDataForSpotMarket(outMarketIndex);
+		const outOraclePrice = outOraclePriceData.price;
+		const inStrictOraclePrice = new StrictOraclePrice(inOraclePrice);
+		const outStrictOraclePrice = new StrictOraclePrice(outOraclePrice);
+
 		const inSpotPosition =
 			this.getSpotPosition(inMarketIndex) ||
 			this.getEmptySpotPosition(inMarketIndex);
@@ -2705,11 +2662,17 @@ export class User {
 		const {
 			totalAssetValue: inTotalAssetValueInitial,
 			totalLiabilityValue: inTotalLiabilityValueInitial,
-		} = this.calculateSpotPositionLeverageContribution(inSpotPosition);
+		} = this.calculateSpotPositionLeverageContribution(
+			inSpotPosition,
+			inStrictOraclePrice
+		);
 		const {
 			totalAssetValue: outTotalAssetValueInitial,
 			totalLiabilityValue: outTotalLiabilityValueInitial,
-		} = this.calculateSpotPositionLeverageContribution(outSpotPosition);
+		} = this.calculateSpotPositionLeverageContribution(
+			outSpotPosition,
+			outStrictOraclePrice
+		);
 
 		const { perpLiabilityValue, perpPnl, spotAssetValue, spotLiabilityValue } =
 			this.getLeverageComponents();
@@ -2728,12 +2691,18 @@ export class User {
 		const {
 			totalAssetValue: inTotalAssetValueAfter,
 			totalLiabilityValue: inTotalLiabilityValueAfter,
-		} = this.calculateSpotPositionLeverageContribution(inPositionAfter);
+		} = this.calculateSpotPositionLeverageContribution(
+			inPositionAfter,
+			inStrictOraclePrice
+		);
 
 		const {
 			totalAssetValue: outTotalAssetValueAfter,
 			totalLiabilityValue: outTotalLiabilityValueAfter,
-		} = this.calculateSpotPositionLeverageContribution(outPositionAfter);
+		} = this.calculateSpotPositionLeverageContribution(
+			outPositionAfter,
+			outStrictOraclePrice
+		);
 
 		const spotAssetValueDelta = inTotalAssetValueAfter
 			.add(outTotalAssetValueAfter)
@@ -2979,6 +2948,7 @@ export class User {
 
 		const assetWeight = calculateAssetWeight(
 			userDepositAmount,
+			oracleData.price,
 			spotMarket,
 			'Initial'
 		);
@@ -3282,6 +3252,8 @@ export class User {
 				spotPosition.marketIndex
 			);
 
+			const strictOraclePrice = new StrictOraclePrice(oraclePriceData.price);
+
 			if (spotPosition.marketIndex === QUOTE_SPOT_MARKET_INDEX) {
 				const tokenAmount = getSignedTokenAmount(
 					getTokenAmount(
@@ -3296,42 +3268,25 @@ export class User {
 				continue;
 			}
 
-			const [worstCaseTokenAmount, worstCaseQuoteTokenAmount] =
-				getWorstCaseTokenAmounts(
-					spotPosition,
-					spotMarketAccount,
-					oraclePriceData
-				);
-
-			netQuoteValue = netQuoteValue.add(worstCaseQuoteTokenAmount);
-
-			const baseAssetValue = getTokenValue(
-				worstCaseTokenAmount.abs(),
-				spotMarketAccount.decimals,
-				oraclePriceData
+			const {
+				tokenAmount: worstCaseTokenAmount,
+				tokenValue: tokenValue,
+				weight,
+				weightedTokenValue: weightedTokenValue,
+				ordersValue: ordersValue,
+			} = getWorstCaseTokenAmounts(
+				spotPosition,
+				spotMarketAccount,
+				strictOraclePrice,
+				marginCategory
 			);
-			const isLiability = isVariant(spotPosition.balanceType, 'borrow');
 
-			let weight;
-			if (isLiability) {
-				weight = calculateLiabilityWeight(
-					worstCaseTokenAmount.abs(),
-					spotMarketAccount,
-					marginCategory
-				);
-			} else {
-				weight = calculateAssetWeight(
-					worstCaseTokenAmount,
-					spotMarketAccount,
-					marginCategory
-				);
-			}
+			netQuoteValue = netQuoteValue.add(ordersValue);
 
-			const weightedValue = baseAssetValue
-				.mul(weight)
-				.div(SPOT_MARKET_WEIGHT_PRECISION);
+			const baseAssetValue = tokenValue.abs();
+			const weightedValue = weightedTokenValue.abs();
 
-			if (isLiability) {
+			if (weightedTokenValue.lt(ZERO)) {
 				healthComponents.borrows.push({
 					marketIndex: spotMarketAccount.marketIndex,
 					size: worstCaseTokenAmount,
@@ -3357,38 +3312,26 @@ export class User {
 			);
 
 			const baseAssetValue = getTokenValue(
-				netQuoteValue.abs(),
+				netQuoteValue,
 				spotMarketAccount.decimals,
 				oraclePriceData
 			);
-			const isLiability = netQuoteValue.lt(ZERO);
 
-			let weight;
-			if (isLiability) {
-				weight = calculateLiabilityWeight(
-					netQuoteValue.abs(),
-					spotMarketAccount,
-					marginCategory
-				);
-			} else {
-				weight = calculateAssetWeight(
-					netQuoteValue,
-					spotMarketAccount,
-					marginCategory
-				);
-			}
+			const { weight, weightedTokenValue } = calculateWeightedTokenValue(
+				netQuoteValue,
+				baseAssetValue,
+				oraclePriceData,
+				spotMarketAccount,
+				marginCategory
+			);
 
-			const weightedValue = baseAssetValue
-				.mul(weight)
-				.div(SPOT_MARKET_WEIGHT_PRECISION);
-
-			if (isLiability) {
+			if (netQuoteValue.lt(ZERO)) {
 				healthComponents.borrows.push({
 					marketIndex: spotMarketAccount.marketIndex,
 					size: netQuoteValue,
-					value: baseAssetValue,
+					value: baseAssetValue.abs(),
 					weight: weight,
-					weightedValue: weightedValue,
+					weightedValue: weightedTokenValue.abs(),
 				});
 			} else {
 				healthComponents.deposits.push({
@@ -3396,7 +3339,7 @@ export class User {
 					size: netQuoteValue,
 					value: baseAssetValue,
 					weight: weight,
-					weightedValue: weightedValue,
+					weightedValue: weightedTokenValue,
 				});
 			}
 		}